requirements.txt
streamlit
pandas
plotly
python-dotenv
binance-connector
tradingview-ta
pymongo
